feat(buffer): upgrade circular buffer to support 128KB+ sizes (#88)

+++ firmware/src/Util/CircularBuffer.h
@@ -34,18 +34,18 @@ typedef struct s_CircularBuf
 {
     uint8_t*    insertPtr;
     uint8_t*    removePtr;
-    uint16_t    totalBytes;
+    uint32_t    totalBytes;
     uint8_t*    buf_ptr;
-    uint16_t    buf_size;
-    int        (*process_callback)(uint8_t*, uint16_t);
+    uint32_t    buf_size;
+    int        (*process_callback)(uint8_t*, uint32_t);
 }CircularBuf_t;
 
 
-void     CircularBuf_Init(CircularBuf_t*, int (*fp)(uint8_t*,uint16_t), uint16_t);
-uint16_t CircularBuf_AddBytes(CircularBuf_t*, uint8_t*, uint16_t);
-uint16_t CircularBuf_NumBytesAvailable(CircularBuf_t*);
-uint16_t CircularBuf_NumBytesFree(CircularBuf_t*);
-uint16_t CircularBuf_ProcessBytes(CircularBuf_t*,uint8_t*, uint16_t,int*);
+void     CircularBuf_Init(CircularBuf_t*, int (*fp)(uint8_t*,uint32_t), uint32_t);
+uint32_t CircularBuf_AddBytes(CircularBuf_t*, uint8_t*, uint32_t);
+uint32_t CircularBuf_NumBytesAvailable(CircularBuf_t*);
+uint32_t CircularBuf_NumBytesFree(CircularBuf_t*);
+uint32_t CircularBuf_ProcessBytes(CircularBuf_t*,uint8_t*, uint32_t,int*);
 void CircularBuf_Reset(CircularBuf_t* cirbuf);
     /* Provide C++ Compatibility */
 #ifdef __cplusplus
